Express description catamorphically.
